Saved model parameters.
Directories with saved model parameters after training the models used in this project. Parameters such as history of performance, loss of the training and validation data during each iteration is saved. The model weights of the best model (least validation loss) is saved during training and the same weights are then used for predictions